refactor: 이미지 확장자 webp로 변경
