+++ src/components/common/myProjectList/ProjectItem.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import COLORS from 'assets/styles/colors';
 import { EditType } from 'types/write/writeType';
-import defaultThumbnail from 'assets/images/thumbnail_small.jpg';
+import defaultThumbnail from 'assets/images/thumbnail_small.webp';
 import ProjectItemMembers from './ProjectItemMembers';
 import { concatSkills } from 'utils/skills';
 import UserStacks from 'components/publicProfile/UserStacks';
