+++ src/components/projectDetail/TitleThumbnailArea.tsx
@@ -6,7 +6,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { deleteProject, updateRecruiting } from 'apis/postDetail';
 import { useAuth, useModal } from 'hooks';
 import { useEffect } from 'react';
-import defaultThumbnail from 'assets/images/thumbnail_big.jpg';
+import defaultThumbnail from 'assets/images/thumbnail_big.webp';
 import {
   amplitudeToNoneButtonClick,
   amplitudeNeedToButtonClick,
